requirements.txt
# requirements.txt

# Core dependencies
pandas>=2.2.0
openpyxl>=3.1.2
python-dotenv>=1.0.0
pyyaml>=6.0.1

# AI dependencies
google-generativeai>=0.3.1

# Development dependencies
pytest>=7.4.0
black>=23.12.1
flake8>=7.0.0
mypy>=1.8.0
